Move multigrid example scripts into pyro/ (#141)
Fixes #140.
