--- examples/__init__.py
@@ -1,1 +1,0 @@
-# example scripts for the solvers
