+++ pyro/multigrid/examples/mg_test_vc_periodic.py
@@ -1,19 +1,17 @@
 #!/usr/bin/env python3
 
-"""Test the general MG solver with a variable coeffcient Helmholtz
-problem.  This ensures we didn't screw up the base functionality here.
+"""
+Test the variable-coefficient MG solver with periodic data.
 
 Here we solve::
 
-   alpha phi + div . ( beta grad phi ) = f
+   div . ( alpha grad phi ) = f
 
 with::
 
-   alpha = 1.0
-   beta = 2.0 + cos(2.0*pi*x)*cos(2.0*pi*y)
+   alpha = 2.0 + cos(2.0*pi*x)*cos(2.0*pi*y)
 
-   f = (-16.0*pi**2*cos(2*pi*x)*cos(2*pi*y) - 16.0*pi**2 + 1.0)*sin(2*pi*x)*sin(2*pi*y)
-
+   f = -16.0*pi**2*(cos(2*pi*x)*cos(2*pi*y) + 1)*sin(2*pi*x)*sin(2*pi*y)
 
 This has the exact solution::
 
@@ -21,25 +19,22 @@ This has the exact solution::
 
 on [0,1] x [0,1]
 
-We use Dirichlet BCs on phi.  For beta, we do not have to impose the
+We use Dirichlet BCs on phi.  For alpha, we do not have to impose the
 same BCs, since that may represent a different physical quantity.
-Here we take beta to have Neumann BCs.  (Dirichlet BCs for beta will
+Here we take alpha to have Neumann BCs.  (Dirichlet BCs for alpha will
 force it to 0 on the boundary, which is not correct here)
 
 """
 
 
-import os
-
 import matplotlib.pyplot as plt
 import numpy as np
 
-from pyro.util import compare
 import pyro.mesh.boundary as bnd
-import pyro.mesh.patch as patch
-import pyro.multigrid.general_MG as MG
+import pyro.multigrid.variable_coeff_MG as MG
 import pyro.util.io_pyro as io
-import pyro.util.msg as msg
+from pyro.mesh import patch
+from pyro.util import compare, msg
 
 
 # the analytic solution
@@ -49,31 +44,19 @@ def true(x, y):
 
 # the coefficients
 def alpha(x, y):
-    return np.ones_like(x)
-
-
-def beta(x, y):
     return 2.0 + np.cos(2.0*np.pi*x)*np.cos(2.0*np.pi*y)
-
-
-def gamma_x(x, y):
-    return np.zeros_like(x)
-
-
-def gamma_y(x, y):
-    return np.zeros_like(x)
 
 
 # the righthand side
 def f(x, y):
-    return (-16.0*np.pi**2*np.cos(2*np.pi*x)*np.cos(2*np.pi*y) -
-            16.0*np.pi**2 + 1.0)*np.sin(2*np.pi*x)*np.sin(2*np.pi*y)
+    return -16.0*np.pi**2*(np.cos(2*np.pi*x)*np.cos(2*np.pi*y) + 1) * \
+        np.sin(2*np.pi*x)*np.sin(2*np.pi*y)
 
 
-def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench_dir="tests/",
-                                   make_plot=False, verbose=1, rtol=1.e-12):
+def test_vc_poisson_periodic(N, store_bench=False, comp_bench=False, bench_dir="tests/",
+                             make_plot=False, verbose=1, rtol=1.e-12):
     """
-    test the general MG solver.  The return value
+    test the variable-coefficient MG solver.  The return value
     here is the error compared to the exact solution, UNLESS
     comp_bench=True, in which case the return value is the
     error compared to the stored benchmark
@@ -86,32 +69,24 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
     # create the coefficient variable
     g = patch.Grid2d(nx, ny, ng=1)
     d = patch.CellCenterData2d(g)
-    bc_c = bnd.BC(xlb="neumann", xrb="neumann",
-                  ylb="neumann", yrb="neumann")
-    d.register_var("alpha", bc_c)
-    d.register_var("beta", bc_c)
-    d.register_var("gamma_x", bc_c)
-    d.register_var("gamma_y", bc_c)
+    bc_c = bnd.BC(xlb="periodic", xrb="periodic",
+                  ylb="periodic", yrb="periodic")
+    d.register_var("c", bc_c)
     d.create()
 
-    a = d.get_var("alpha")
-    a[:, :] = alpha(g.x2d, g.y2d)
+    c = d.get_var("c")
+    c[:, :] = alpha(g.x2d, g.y2d)
 
-    b = d.get_var("beta")
-    b[:, :] = beta(g.x2d, g.y2d)
-
-    gx = d.get_var("gamma_x")
-    gx[:, :] = gamma_x(g.x2d, g.y2d)
-
-    gy = d.get_var("gamma_y")
-    gy[:, :] = gamma_y(g.x2d, g.y2d)
+    # check whether the RHS sums to zero (necessary for periodic data)
+    rhs = f(g.x2d, g.y2d)
+    print(f"rhs sum: {np.sum(rhs[g.ilo:g.ihi+1, g.jlo:g.jhi+1])}")
 
     # create the multigrid object
-    a = MG.GeneralMG2d(nx, ny,
-                       xl_BC_type="dirichlet", yl_BC_type="dirichlet",
-                       xr_BC_type="dirichlet", yr_BC_type="dirichlet",
-                       coeffs=d,
-                       verbose=verbose, vis=0, true_function=true)
+    a = MG.VarCoeffCCMG2d(nx, ny,
+                          xl_BC_type="periodic", yl_BC_type="periodic",
+                          xr_BC_type="periodic", yr_BC_type="periodic",
+                          coeffs=c, coeffs_bc=bc_c,
+                          verbose=verbose, vis=0, true_function=true)
 
     # initialize the solution to 0
     a.init_zeros()
@@ -124,14 +99,20 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
     a.solve(rtol=1.e-11)
 
     # alternately, we can just use smoothing by uncommenting the following
-    # a.smooth(a.nlevels-1,50000)
+    # a.smooth(a.nlevels-1,10000)
 
     # get the solution
     v = a.get_solution()
 
-    # compute the error from the analytic solution
+    # get the true solution
     b = true(a.x2d, a.y2d)
-    e = v - b
+
+    # compute the error from the analytic solution -- note that with
+    # periodic BCs all around, there is nothing to normalize the
+    # solution.  We subtract off the average of phi from the MG
+    # solution (we do the same for the true solution to put them on
+    # the same footing)
+    e = v - np.sum(v.v())/(nx*ny) - (b - np.sum(b[a.ilo:a.ihi+1, a.jlo:a.jhi+1])/(nx*ny))
 
     enorm = e.norm()
     print(" L2 error from true solution = %g\n rel. err from previous cycle = %g\n num. cycles = %d" %
@@ -169,10 +150,10 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
 
         plt.tight_layout()
 
-        plt.savefig("mg_general_alphabeta_only_test.png")
+        plt.savefig("mg_vc_periodic_test.png")
 
     # store the output for later comparison
-    bench = "mg_general_poisson_dirichlet"
+    bench = "mg_vc_poisson_periodic"
 
     my_data = a.get_solution_object()
 
@@ -182,7 +163,7 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
     # do we do a comparison?
     if comp_bench:
         compare_file = f"{bench_dir}/{bench}"
-        msg.warning("comparing to: %s " % (compare_file))
+        msg.warning(f"comparing to {compare_file}")
         bench = io.read(compare_file)
 
         result = compare.compare(my_data, bench, rtol)
@@ -190,7 +171,7 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
         if result == 0:
             msg.success(f"results match benchmark to within relative tolerance of {rtol}\n")
         else:
-            msg.warning("ERROR: " + compare.errors[result] + "\n")
+            msg.warning(f"ERROR: {compare.errors[result]}\n")
 
         return result
 
@@ -200,7 +181,7 @@ def test_general_poisson_dirichlet(N, store_bench=False, comp_bench=False, bench
 
 if __name__ == "__main__":
 
-    N = [16, 32, 64]
+    N = [16, 32, 64, 128, 256, 512]
     err = []
 
     plot = False
@@ -211,9 +192,8 @@ if __name__ == "__main__":
         if nx == max(N):
             plot = True
 
-        enorm = test_general_poisson_dirichlet(nx, make_plot=plot,
-                                               store_bench=store, comp_bench=do_compare)
-
+        enorm = test_vc_poisson_periodic(nx, make_plot=plot,
+                                         store_bench=store, comp_bench=do_compare)
         err.append(enorm)
 
     # plot the convergence
@@ -232,4 +212,4 @@ if __name__ == "__main__":
 
     plt.tight_layout()
 
-    plt.savefig("mg_general_alphabeta_only_converge.png")
+    plt.savefig("mg_vc_periodic_converge.png")
