+++ pyro/test.py
@@ -7,8 +7,9 @@ import os
 import sys
 
 import pyro.pyro_sim as pyro
-from examples.multigrid import (mg_test_general_inhomogeneous, mg_test_simple,
-                                mg_test_vc_dirichlet, mg_test_vc_periodic)
+from pyro.multigrid.examples import (mg_test_general_inhomogeneous,
+                                     mg_test_simple, mg_test_vc_dirichlet,
+                                     mg_test_vc_periodic)
 
 
 class PyroTest:
